update project structure and routes
