+++ src/controller/register-controller.js
@@ -2,7 +2,6 @@ import { registerUserEmail } from '../model/register-model.js';
 
 export default () => {
   const btnRegister = document.getElementById('btn-register');
-  console.log(btnRegister);
   btnRegister.addEventListener('click', (event) => {
     event.preventDefault();
     const email = document.querySelector('#email-register');
