+++ src/routes/components.js
@@ -1,0 +1,9 @@
+import viewRegister from '../view/register-view.js';
+import viewLogin from '../view/login-view.js';
+import viewHome from '../view/home-view.js';
+
+export const components = {
+  register: viewRegister,
+  login: viewLogin,
+  home: viewHome,
+};
