+++ src/routes/router.js
@@ -1,6 +1,6 @@
-import { components } from '../view/index.js';
-import registerController from './register-controller.js';
-import { signInEmailEvent } from './login-controller.js';
+import { components } from './components.js';
+import registerController from '../controller/register-controller.js';
+import { signInEmailEvent } from '../controller/login-controller.js';
 
 export const changeView = (hash) => {
   const container = document.querySelector('#container');
@@ -19,7 +19,18 @@ export const changeView = (hash) => {
     case '#/home':
       container.appendChild(components.home());
       break;
+    case '#/signOut':
+      firebase
+        .auth()
+        .signOut()
+        .then(() => {
+          console.log('sesion cerrada');
+        }).catch((error) => {
+          console.log(error);
+        });
+      break;
     default:
+      container.appendChild(components.login());
       break;
   }
 };
