+++ src/view/home-view.js
@@ -1,7 +1,7 @@
 export default () => {
   const homeView = `
   <p>Bienvenidx</p>
-    <button id="btn-signout" class="btn-general" type="button">SALIR</button>`;
+  <a href="#/">Cerrar Sesión</a>`;
   const divELem = document.createElement('div');
   divELem.innerHTML = homeView;
   return divELem;
