--- src/view/index.js
@@ -1,9 +1,0 @@
-import viewRegister from './register-view.js';
-import viewLogin from './login-view.js';
-import viewHome from './home-view.js';
-
-export const components = {
-  register: viewRegister,
-  login: viewLogin,
-  home: viewHome,
-};
